Split to base file
